New systems
New systems for gravity and obj
new system for the selection of html or compiled
moved the html stuff to src/html
updates to the ui
added the list function

+++ src/simloader.rs
@@ -14,13 +14,15 @@ pub fn loadsim(file:&str) -> Vec<obj::obj>{
             // then creates the function
             let loadsimplan: obj::obj = obj::obj{
                 name: data[0].to_string(),
-                x:  data[1].parse::<f32>().expect("parse error") as f32,
-                y: data[2].parse::<f32>().expect("parse error2") as f32,
-                mass: data[3].parse::<f32>().expect("parse error3") as f32,
-                size: data[4].parse::<f32>().expect("parse error4") as f32,
-                velx: data[5].parse::<f32>().expect("parse error5") as f32,
-                vely: data[6].parse::<f32>().expect("parse error6") as f32,
-                bounce: data[7].parse::<f32>().expect("parse error7") as f32,
+                x:  data[1].parse::<f64>().expect("parse error") as f64,
+                y: data[2].parse::<f64>().expect("parse error2") as f64,
+                mass: data[3].parse::<f64>().expect("parse error3") as f64,
+                size: data[4].parse::<f64>().expect("parse error4") as f64,
+                velx: data[5].parse::<f64>().expect("parse error5") as f64,
+                vely: data[6].parse::<f64>().expect("parse error6") as f64,
+                bounce: data[7].parse::<f64>().expect("parse error7") as f64,
+            
+
             };
             returndata.push(loadsimplan);
         }else{
